Increase max length for ingredient properties
Updated the `Name` and `Quantity` properties in the
`IngredientEntity` class to a maximum length of 100 characters.
The `Name` property in the `BlockIngredientListEntity` class
has also been updated to match. A new migration file
`20250222210737_UpdateLengthOfAllIngredientsFields`
has been created to apply these changes to the database schema,
and the `ApplicationDbContextModelSnapshot.cs` file
has been updated accordingly.